Move toolchain base class into another file
