+++ tools/toolchains/gcc.py
@@ -21,7 +21,7 @@ from distutils.spawn import find_executable
 from distutils.version import LooseVersion
 
 from tools.targets import CORE_ARCH
-from tools.toolchains import mbedToolchain, TOOLCHAIN_PATHS
+from tools.toolchains.mbed_toolchain import mbedToolchain, TOOLCHAIN_PATHS
 from tools.utils import run_cmd, NotSupportedException
 
 class GCC(mbedToolchain):
